\name{SpGPCW-package}
\alias{SpGPCW-package}
\alias{SpGPCW}
\docType{package}
\title{
\packageTitle{SpGPCW}
}
\description{
\packageDescription{SpGPCW}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SpGPCW}
\packageIndices{SpGPCW}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{SpGPCW}

Maintainer: \packageMaintainer{SpGPCW}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
